code refactor - StatusModal used in StatusCard component
