+++ components/Admin/Details/DetailsPlate/DetailsPlate.jsx
@@ -4,7 +4,6 @@ import BackgroundCanvas from "../../Shared/BackgroundCanvas/BackgroundCanvas";
 import TextParagraph from "../../Shared/Typography/TextParagraph";
 import TextBold from "../../Shared/Typography/TextBold";
 import { dateFormat } from "@/utils/serialize";
-import StatusMessage from "../StatusMessage/StatusMessage";
 
 export default function DetailsPlate({
   id,
@@ -13,25 +12,8 @@ export default function DetailsPlate({
   location,
   description,
 }) {
-  const [openModal, setOpenModal] = useState(false);
-  console.log("openModal", openModal);
-
-  function close(e) {
-    if (e.target.id === "overlay") {
-      setOpenModal(false);
-    }
-  }
-
   return (
     <BackgroundCanvas className={styles.detailsMain}>
-      <button onClick={() => setOpenModal(true)}>Modal</button>
-      <div
-        id="overlay"
-        className={`${openModal === true ? styles.overlay : undefined}`}
-        onClick={openModal ? close : null}
-      >
-        <StatusMessage open={openModal} onClose={() => setOpenModal(false)} />
-      </div>{" "}
       <div className={styles.top}>
         <div className={styles.id}>
           <div className={`${styles.idInner} ${styles.idSpacing}`}>
