+++ components/Admin/Details/DetailsPlate/DetailsPlate.module.css
@@ -1,13 +1,3 @@
-.overlay {
-  background-color: rgba(255, 255, 255, 0.9);
-  position: fixed;
-  inset: 0;
-  z-index: 30;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .id {
   display: flex;
   flex-direction: column;
