+++ components/Admin/Details/StatusCard/StatusCard.module.css
@@ -15,6 +15,18 @@
   }
 }
 
+.overlay {
+  background-color: rgba(255, 255, 255, 0.9);
+  position: fixed;
+  margin-top: -20rem;
+
+  inset: 0;
+  z-index: 30;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 .statusCardContainer {
   display: flex;
   justify-content: space-between;
