--- components/Admin/Details/StatusMessage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./StatusMessage";
